Fix for task 029
